jitter_timestamp with variable name change
JITTER command in recipe generates an error that points to this file.
jitter_timestamp was called in line 121 with "item=dicom" but the
function itself now asks for parameter name "dicom" so changed line 121
to dicom=dicom

Also the code for jitter adding datetime together with simple + in line
160 feels unusual for date operation.  Added new code to utilize
datetime package to do the jitter and updates the dicom field.

In dicom/utils/action.py there is a get_timestamp func which is imported
by actions.py and seems to deal with jitter but looks to be heavier than
necessary since JITTER operates in +/- days only based on Documentation.

Neither current nor old jitter_timestamp uses the get_timestamp
function.