adding chunk of work to allow for parsing of private dicom tags
this means small updates to most functions to have special handling in
the case of being provided a string (field name) versus a tag (BaseTag)
that will produce a DataElement for dicom.get(field). This will need
robust testing to ensure that we dont need to be more specific about
some if these if/else (e.g., checking for a type instead). Additionally,
Ive added the dicom_dir() function under fields that will serve the
same purpose as dicom.dir() but also provide the keys to private tags,
which are type pydicom.tag.BaseTag

Signed-off-by: vsoch <vsochat@stanford.edu>