Add support to parse private tags (#125)
* Adding support for private tags groups to be parsed

Signed-off-by: vsoch <vsochat@stanford.edu>

* adding chunk of work to allow for parsing of private dicom tags

this means small updates to most functions to have special handling in
the case of being provided a string (field name) versus a tag (BaseTag)
that will produce a DataElement for dicom.get(field). This will need
robust testing to ensure that we dont need to be more specific about
some if these if/else (e.g., checking for a type instead). Additionally,
Ive added the dicom_dir() function under fields that will serve the
same purpose as dicom.dir() but also provide the keys to private tags,
which are type pydicom.tag.BaseTag

Signed-off-by: vsoch <vsochat@stanford.edu>

* expanded expressions cannot include private tags, as they are not strings

Signed-off-by: vsoch <vsochat@stanford.edu>

* New Sample Image (#126)

Adding new sample image created from a Head CT.
Updated number of expected test files.

* small fix to allow for parsing private tags in contenders

currently, if you parse through a list of fields with a private tag,
for pydicom 1.2.1 you will get an error about comparison of a non tag
with a tag. This doesnt happen with 1.4.1, however 1.2.1 needs to still
be supported

Signed-off-by: vsoch <vsochat@stanford.edu>

Co-authored-by: wetzelj <wetzelj@ccf.org>