adding checks to see if we have a file meta field
there is currently a bug that when we add (or otherwise manipulate a field) we assume
retrieving the field from the main dicom. We actually need to check if the field is from
the file meta, and change behavior based on that. So these changes should fix a bug of
running the example and trying to save the dicom.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>