Add support to read and manipulate file meta (#184)
* This will add support to read and manipulate file meta

Remaining questions:
1. Do file meta fields overlap with fields in the dicom? If so, we need to be checking each field if it is a filemeta or not before replace. As implemented now, the fields are parsing over equivalently and it is assumed that file meta fields do not appear in the dicom and vice versa.
2. Do we need to make any special changes to the file meta, e.g., perhaps the size? I have never manipulated it before, and I am hoping that pydicom save handles these changes, but if not we should do it manually.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* updating file meta parsing to not be "special"

instead of having a special section, we can include file-meta parsing alongside the
header parsing, as the names are unique. The one tiny detail is that we want to be
careful with fields that should not be touched, so those are now added to the
set of "skip." This set also includes other like PixelData that was not previously
being loaded, which I would consider a bug (that should now be fixed)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* adding checks to see if we have a file meta field

there is currently a bug that when we add (or otherwise manipulate a field) we assume
retrieving the field from the main dicom. We actually need to check if the field is from
the file meta, and change behavior based on that. So these changes should fix a bug of
running the example and trying to save the dicom.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* small bugfix in test and checking element tag

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

Co-authored-by: vsoch <vsoch@users.noreply.github.com>