LGTM recommendation: Unused import
Import of 'Sequence' is not used.
Import of 'remove_sequences' is not used.
Import of 'os' is not used.
Import of 'read_json' is not used.
Import of 'DeidRecipe' is not used.
Import of 'get_private' is not used.
Import of 'remove_sequences' is not used.
Import of 'extract_fields_list' is not used.
Import of 'extract_values_list' is not used.
Import of 'get_fields' is not used.
Import of 'DataElement' is not used.
Import of 'operator' is not used.
Import of 'get_installdir' is not used.
Import of 'os' is not used.
mport of 'read_file' is not used.
Import of 'os' is not used.