Adding ability for deid to support deid-provided functions (#208)
* adding ability for deid to support deid-provided functions

as of this version, a user can specify a value as a deid_func:<name> meaning that we use
a deid provided function. This is useful for providing some general uid and a customized jitter
function, along with other functions that users might want to add. With this change I have
provided docs and a contributing section for how to do this, along with running pyflakes on
the tests
Signed-off-by: vsoch <vsoch@users.noreply.github.com>