Fix assignment of deided attributes to correct nested tags
In cases where tags are deeply nested within sequences, the output of
deid functions were incorrectly being assigned to the top-level
tags with matching names. For example, a nested SeriesInstanceUID within
a ReferencedSeriesSequence was overwriting the top-level
SeriesInstanceUID value.

This was resolved by traversing the tree of tags and setting the nested
element, rather than the top-level element.