Use lookup tables to reduce field parsing
The `get_fields_with_lookup` function was added to augment the
`get_fields` function with a lookup table that allows quick
identification of exact tag matches.

This optimization significantly reduced the amount of time spent in the
exact matching stage of `expand_field_expression`. For top-level DICOM
tag searches, the search in "Case 2" would call `name_contains` on every
single tag in the DICOM dataset. With lookup tables, we can look up a
contender field based on the values for which we're checking exact
matches-- this becomes a key lookup problem rather than searching all
fields for a matching identifier.