Improve speed of header deid with lookup tables and caching (#289)
* Pre-compile regexes and use string comparison where possible

During profiling, it was identified that repeated regex lookups inside
loops were taking a significant amount of time. To reduce this, regex
expressions were pre-compiled outside of loops, and plain string
comparison was used when possible to sidestep the performance overhead
of regex matching for simple things like string equality comparison.

* Use lookup tables to reduce field parsing

The `get_fields_with_lookup` function was added to augment the
`get_fields` function with a lookup table that allows quick
identification of exact tag matches.

This optimization significantly reduced the amount of time spent in the
exact matching stage of `expand_field_expression`. For top-level DICOM
tag searches, the search in "Case 2" would call `name_contains` on every
single tag in the DICOM dataset. With lookup tables, we can look up a
contender field based on the values for which we're checking exact
matches-- this becomes a key lookup problem rather than searching all
fields for a matching identifier.

* Add cache layer around field retrieval

* Fix test failures

* Use isinstance for type comparison

* Add docstrings to expander functions

* Use a wrapper class for field lookup tables

* Respond to PR feedback

* Fix formatting

* Remove case-sensitivity of fields in lookup tables

* Remove mentions of whole_string kwarg in docs

* Move scope of hash function override to local function

* Switch to context manager for cleaner override